requirements.txt
pygame>=2.0.0
gpiozero>=1.6.0
moviepy>=1.0.3
numpy>=1.19.0
openai>=1.0.0
requests>=2.25.0
pillow>=8.0.0
